Factor out mono literal and pattern into smaller crates
